fix: taxonomy alignment and error message
