+++ src/taxonomy/data/apiHooks.ts
@@ -229,18 +229,13 @@ export const useSubTags = (taxonomyId: number, parentTagValue: string) =>
 
 export const useCreateTag = (taxonomyId: number) => {
   const queryClient = useQueryClient();
-  const intl = useIntl();
 
   return useMutation({
-    mutationFn: async ({ value, parentTagValue }: { value: string; parentTagValue?: string; }) => {
-      try {
-        await getAuthenticatedHttpClient().post(
-          apiUrls.createTag(taxonomyId),
-          { tag: value, parent_tag_value: parentTagValue },
-        );
-      } catch (err) {
-        throw new Error(getApiErrorMessage(err, intl));
-      }
+    mutationFn: async ({ value, parentTagValue }: { value: string, parentTagValue?: string }) => {
+      await getAuthenticatedHttpClient().post(
+        apiUrls.createTag(taxonomyId),
+        { tag: value, parent_tag_value: parentTagValue },
+      );
     },
     onSuccess: () => {
       queryClient.invalidateQueries({
@@ -260,15 +255,11 @@ export const useUpdateTag = (taxonomyId: number) => {
   const queryClient = useQueryClient();
 
   return useMutation({
-    mutationFn: async ({ value, originalValue }: { value: string; originalValue: string; }) => {
-      try {
-        await getAuthenticatedHttpClient().patch(
-          apiUrls.updateTag(taxonomyId),
-          { tag: originalValue, updated_tag_value: value },
-        );
-      } catch (err) {
-        throw new Error(getApiErrorMessage(err));
-      }
+    mutationFn: async ({ value, originalValue }: { value: string, originalValue: string }) => {
+      await getAuthenticatedHttpClient().patch(
+        apiUrls.updateTag(taxonomyId),
+        { tag: originalValue, updated_tag_value: value },
+      );
     },
     onSuccess: () => {
       queryClient.invalidateQueries({
